Remove Vertex IDs, related mappings etc.
