--- tests/Set/VerticesMapTest.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace Graphp\Graph\Tests\Set;
-
-use Graphp\Graph\Set\VerticesMap;
-
-class VerticesMapTest extends BaseVerticesTest
-{
-    protected function createVertices(array $vertices)
-    {
-        return new VerticesMap($vertices);
-    }
-}
